+++ src/ts/spatialnavigation/types.ts
@@ -6,7 +6,7 @@ export type AnyComponent = Component<ComponentConfig>;
 export type AnyContainer = Container<ContainerConfig>;
 export type Callback<T> = (data: T, target: AnyComponent, preventDefault: () => void) => boolean | void;
 export type NavigationCallback = Callback<Direction>;
-export type NavigationNoTargetCallback = (direction: Direction) => void;
+export type AfterNavigationCallback = (direction: Direction, target?: AnyComponent) => void;
 export type ActionCallback = Callback<Action>;
 export type KeyMap = {
   [keyCode: number]: Action | Direction;