+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `NavigationGroup.afterNavigation` to detect after the directional navgiation finished
+
 ## [4.11.1] - 2026-04-23
 
 ### Fixed